+++ R/estimate_vpd_from_dew.R
@@ -1,15 +1,15 @@
 #' Estimate VPD from daily minimum temperature
 #'
-#' of the data in the class funciton using \code{\link{estimate_vpd_from_dew}}.
+#' of the data in the class function using \code{\link{estimate_vpd_from_dew}}.
 #'
 #' @param ... further arguments to \code{\link{estimate_vpd_from_dew}}
 #'
 #' @return side effect of updated column VPDfromDew in class
 #' @export
-sEddyProc_estimate_vpd_from_dew <- function(...){
+sEddyProc_sFillVPDFromDew <- function(...){
   VPDfromDew <- cbind(.self$sDATA, select(.self$sTEMP, -.data$sDateTime)) %>%
      mutate(DateTime = .data$sDateTime) %>%
-     REddyProc::estimate_vpd_from_dew(...)
+     estimate_vpd_from_dew(...)
   .self$sTEMP$VPDfromDew <- VPDfromDew
   if (!("VPD_f" %in% names(.self$sTEMP))) {
     warning(
@@ -21,8 +21,8 @@ sEddyProc_estimate_vpd_from_dew <- function(...){
   .self$sTEMP$VPD_f[iNA] <- VPDfromDew[iNA]
   invisible(.self)
 }
-sEddyProc$methods(estimate_vpd_from_dew =
-                    sEddyProc_estimate_vpd_from_dew)
+sEddyProc$methods(sFillVPDFromDew =
+                    sEddyProc_sFillVPDFromDew)
 
 
 #' Estimate VPD from assuming dewpoint at daily minimum temperature
@@ -72,14 +72,16 @@ estimate_vpd_from_dew <- function(df, pNonMissing = 0.1){
       TmaxOftheDay = suppressWarnings(max(.data$Tair_f, na.rm = TRUE)),
       TRangeDay = .data$TmaxOftheDay - .data$TminOftheDay,
       EminDew = Esat.slope(.data$TminOftheDay)$Esat*10,
-      VPDfromDew = Esat.slope(.data$Tair)$Esat*10 - .data$EminDew # here use only Tair
+      VPDfromDew = Esat.slope(.data$Tair_f)$Esat*10 - .data$EminDew 
     ) %>%
     ungroup()
-  # ok to use Tair_f here because VPDfromDew is NA when Tair is NA
-  lm1 <- lm(VPD ~ 0 + VPDfromDew * Tair_f * hourOfDay *  TminOftheDay * TRangeDay, df_f)
+  # VPDfromDew is computed using Tair_f, but for the correction only use
+  # those cases where original Tair is finite to avoid confounding issues
+  lm1 <- lm(VPD ~ 0 + VPDfromDew * Tair_f * hourOfDay *  TminOftheDay * TRangeDay, 
+            filter(df_f, is.finite(.data$Tair)))
   # create column in original data.frame (do not return the intermediate vars)
   # the order in df and df_f should not have changed with grouping/ungrouping
-  VPDfromDew = predict(lm1, df_f)
+  VPDfromDew = pmax(0,predict(lm1, df_f))
 }
 
 getCumDay <- function(